remove old managers, adjust some tests (#163)
