+++ eogrow/core/area/__init__.py
@@ -2,6 +2,6 @@
 A subfolder for area management
 """
 
-from .batch import BatchAreaManager, NewBatchAreaManager
-from .custom_grid import CustomGridAreaManager, NewCustomGridAreaManager
-from .utm import NewUtmZoneAreaManager, UtmZoneAreaManager
+from .batch import BatchAreaManager
+from .custom_grid import CustomGridAreaManager
+from .utm import UtmZoneAreaManager
