+++ tests/conftest.py
@@ -38,12 +38,13 @@ def temp_folder_fixture():
 
 @pytest.fixture(scope="session", name="test_storage_manager")
 def test_storage_manager_fixture(project_folder):
-    filename = os.path.join(
-        os.path.dirname(os.path.realpath(__file__)), "test_config_files", "other", "local_storage_test.json"
-    )
-    config = interpret_config_from_path(filename)
-    config["storage"]["project_folder"] = project_folder
-    yield StorageManager.from_raw_config(config["storage"])
+    manager_config = {
+        "manager": "eogrow.core.storage.StorageManager",
+        "project_folder": project_folder,
+        "structure": {"temp": "temp", "batch": "tiffs", "eopatches": "path/to/eopatches"},
+    }
+
+    yield StorageManager.from_raw_config(manager_config)
 
 
 @pytest.fixture(name="storage")
