--- tests/test_config_files/other/eopatch_global_config.json
@@ -1,20 +1,0 @@
-{
-  "storage": {
-    "manager": "eogrow.core.storage.StorageManager",
-    "project_folder": "${config_path}/../test_project/"
-  },
-  "area": {
-    "manager": "eogrow.core.area.UtmZoneAreaManager",
-    "area_filename": "test_area.geojson",
-    "area_buffer": 0.005,
-    "patch_size_x": 100,
-    "patch_size_y": 100,
-    "patch_buffer_x": 15,
-    "patch_buffer_y": 15,
-    "offset_x": 0,
-    "offset_y": 0
-  },
-  "eopatch": {
-    "manager": "eogrow.core.eopatch.EOPatchManager"
-  }
-}
