--- tests/test_config_files/other/local_storage_test.json
@@ -1,25 +1,0 @@
-{
-  "area": {
-    "manager": "eogrow.core.area.batch.BatchAreaManager",
-    "area_filename": "test_area_eogrow.geojson",
-    "tiling_grid_id": 1,
-    "resolution": 10,
-    "tile_buffer": 50
-  },
-  "storage": {
-    "manager": "eogrow.core.storage.StorageManager",
-    "project_folder": "${config_path}/../../test_project/",
-    "aws_profile": null,
-    "structure": {
-      "temp": "temp",
-      "batch": "tiffs",
-      "eopatches": "path/to/eopatches"
-    }
-  },
-  "eopatch": {
-    "manager": "eogrow.core.eopatch.BatchTileManager"
-  },
-  "save_logs": true,
-  "show_logs": true,
-  "use_ray": false
-}
